Split integration test cases into different files
