+++ ci/tasks/run-integration-windows.sh
@@ -43,12 +43,12 @@ export BOSH_AZURE_USE_MANAGED_DISKS=${AZURE_USE_MANAGED_DISKS}
 pushd bosh-cpi-src/src/bosh_azure_cpi > /dev/null
   bundle install
 
-  tags="--tag ~heavy_stemcell"
+  tags="--tag ~heavy_stemcell --tag ~migration --tag ~azure_cpi_executable"
   export BOSH_AZURE_USE_MANAGED_DISKS=${AZURE_USE_MANAGED_DISKS}
   if [ "${AZURE_USE_MANAGED_DISKS}" == "true" ]; then
     tags+=" --tag ~unmanaged_disks"
   else
     tags+=" --tag ~availability_zone"
   fi
-  bundle exec rspec spec/integration/lifecycle_spec.rb ${tags}
+  bundle exec rspec spec/integration/ ${tags}
 popd > /dev/null
