+++ ci/tasks/run-integration.sh
@@ -44,32 +44,26 @@ chruby ${RUBY_VERSION}
 pushd bosh-cpi-src/src/bosh_azure_cpi > /dev/null
   bundle install
 
-  tags="--tag ~light_stemcell"
+  tags="--tag ~light_stemcell --tag ~migration"
   export BOSH_AZURE_USE_MANAGED_DISKS=${AZURE_USE_MANAGED_DISKS}
   if [ "${AZURE_USE_MANAGED_DISKS}" == "true" ]; then
     tags+=" --tag ~unmanaged_disks"
   else
     tags+=" --tag ~availability_zone"
   fi
-  bundle exec rspec spec/integration/lifecycle_spec.rb ${tags}
+  bundle exec rspec spec/integration/ ${tags}
 
   # migration: unmanged disk -> managed disk
   # Only run migration test when AZURE_USE_MANAGED_DISKS is set to false initially
   if [ "${AZURE_USE_MANAGED_DISKS}" == "false" ]; then
     unset BOSH_AZURE_USE_MANAGED_DISKS
-    bundle exec rspec spec/integration/managed_disks_migration_spec.rb
+    bundle exec rspec spec/integration/migrations/managed_disks_migration_spec.rb
   fi
 
   # migration: regional resource -> zonal resource
   # Only run migration test when AZURE_USE_MANAGED_DISKS is set to true initially
   if [ "${AZURE_USE_MANAGED_DISKS}" == "true" ]; then
     export BOSH_AZURE_USE_MANAGED_DISKS=true
-    bundle exec rspec spec/integration/availability_zone_migration_spec.rb
-  fi
-  
-  # The azure_cpi test doesn't care whether managed disks are used. Only run it when AZURE_USE_MANAGED_DISKS is set to true
-  if [ "${AZURE_USE_MANAGED_DISKS}" == "true" ]; then
-    export BOSH_AZURE_USE_MANAGED_DISKS=true
-    bundle exec rspec spec/integration/azure_cpi_spec.rb
+    bundle exec rspec spec/integration/migrations/availability_zone_migration_spec.rb
   fi
 popd > /dev/null
